requirements.txt
async-timeout==4.0.2
certifi==2022.6.15
charset-normalizer==2.1.1
Deprecated==1.2.13
idna==3.3
numpy==1.23.2
packaging==21.3
pandas==1.4.4
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
six==1.16.0
spotipy==2.20.0
urllib3==1.26.12
wrapt==1.14.1
